move LocalVerificationStatus
