+++ pallets/creditcoin/src/ocw/task.rs
@@ -1,10 +1,9 @@
 use alloc::vec::Vec;
 use codec::Encode;
 use frame_support::storage::PrefixIterator;
+use sp_runtime::offchain::storage::StorageValueRef;
 
-use super::{
-	Config, LocalVerificationStatus, Pallet, VerificationFailureCause, VerificationResult,
-};
+use super::{Config, Pallet, VerificationFailureCause, VerificationResult};
 
 pub(crate) const TASK_GUARD: &[u8] = b"creditcoin/task/guard";
 
@@ -75,5 +74,35 @@ where
 			},
 			None => None,
 		}
+	}
+}
+
+pub(crate) struct LocalVerificationStatus<'a> {
+	storage_ref: StorageValueRef<'a>,
+	key: &'a [u8],
+}
+
+impl<'a> LocalVerificationStatus<'a> {
+	pub(crate) fn new(storage_key: &'a [u8]) -> Self {
+		Self { storage_ref: StorageValueRef::persistent(storage_key), key: storage_key }
+	}
+
+	pub(crate) fn is_complete(&self) -> bool {
+		match self.storage_ref.get::<()>() {
+			Ok(Some(())) => true,
+			Ok(None) => false,
+			Err(e) => {
+				log::warn!(
+					"Failed to decode offchain storage for {}: {:?}",
+					hex::encode(self.key),
+					e
+				);
+				true
+			},
+		}
+	}
+
+	pub(crate) fn mark_complete(&self) {
+		self.storage_ref.set(&());
 	}
 }
